Moved CustomPool in its own file and moved comments to the class doc string.
